refactor: restructure type files
