+++ src/lambdaFunctions/types/types.ts
@@ -1,22 +1,1 @@
-/**
-* This type will allow Typescript to pick up on pieces of an API request.
-*/
-export type AWSRequest = {
-  body?: string;
-  headers?: {
-    [key: string]: string;
-  },
-  httpMethod?: string,
-  isBase64Encoded?: boolean,
-}
-
-/**
-* This type will allow Typescript to create a response to an API request.
-*/
-export type AWSResponse = {
-  body: string;
-  headers?: {
-    [key: string]: string;
-  },
-  statusCode: number,
-}
+export { AWSRequest, AWSResponse } from "./aws";
